refactor: restructure folder
