+++ src/components/mod.rs
@@ -1,0 +1,2 @@
+pub mod physics;
+pub mod player;
