+++ src/components/player.rs
@@ -1,0 +1,6 @@
+use bevy::prelude::*;
+
+#[derive(Component)]
+pub struct Player {
+    pub movement_speed: f32,
+}
