+++ src/plugins/mod.rs
@@ -1,0 +1,4 @@
+pub mod inputs;
+pub mod physics;
+pub mod player;
+pub mod setup;
