+++ src/systems/setup.rs
@@ -1,0 +1,17 @@
+use bevy::prelude::*;
+
+pub fn spawn_camera(mut commands: Commands) {
+    commands.spawn(Camera2dBundle::default());
+}
+
+pub fn say_hello() {
+    println!(
+        "\
+░▀█▀░█▀▄░█▀█░█▀█    
+░░█░░█▀▄░█░█░█░█    
+░▀▀▀░▀░▀░▀▀▀░▀░▀    
+░█░█░▀█▀░█░░░█▀▄░█▀▀
+░█▄█░░█░░█░░░█░█░▀▀█
+░▀░▀░▀▀▀░▀▀▀░▀▀░░▀▀▀"
+    );
+}
